Java: Construct a Tensor from a handle to the C TF_Tensor object.
This change introduces a package private static method to create
a Java Tensor object from the handle to the native TF_Tensor object.
This will be needed to create the output Tensor objects produced
by graph execution (when the equivalent of a Session and it's run()
method are introduced in Java).

Yet another step in the journey that is #5
Change: 141096422