Create, save and load models using functional API
Python models that contain tf.function is inconvenient to be consumed by Java clients.
This proposal provides an API to
(a) Invoke a tf.function, given the signature name
(b) Retrieve the node name in the graph corresponding to a tf.function

Co-authored-by: Shajan Dasan <sdasan@twitter.com>

Save models as functions (#103)

* Draft: Java API to use tf.function available on SavedModel. (#89)

Python models that contain tf.function is inconvenient to be consumed by Java clients.
This proposal provides an API to
(a) Invoke a tf.function, given the signature name
(b) Retrieve the node name in the graph corresponding to a tf.function

Co-authored-by: Shajan Dasan <sdasan@twitter.com>

* Change API for creating concrete functions and exporting them to a saved model

Co-authored-by: Karl Lessard <klessard@gmail.com>

Rename signature name to key

Print function signature when converting to String

Add method that returns the signature of all functions in a saved model

Add unit tests for python created SavedModel with tf.function